nothing of note

+++ src/test.cpp
@@ -29,7 +29,7 @@ int main(void){
     int8_t out1 = base_convert<int8_t>(0b00011111, 3,2);
     int8_t out2 = base_convert<int8_t>(53, 2,3); 
     std::cout << N << (int)out1 << N << BITS(out1)<< N << (int)out2 << N << BITS(out2) << N;
-    std::cout << log_base<int8_t>(32, 3);
+    std::cout << (int)log_base<int8_t>(32, 3);
 
 
 }
